added SSR for thumbnail support
